Move sequential creation logic to a mixin
